feat: refactor layout structure and implement zoom functionality in Canvas and ToolBar components
