+++ frontend/src/app/metadata.ts
@@ -1,0 +1,10 @@
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: {
+    template: '%s | KIBAKO',
+    default: 'KIBAKO',
+  },
+  description: '気軽にボードゲームを作ろう',
+  icons: [{ rel: 'icon', url: '/favicon.ico' }],
+};
